refactor: centralize workspace metadata and reorganize tests
- Centralize version, edition, and authors in workspace root
- Update all crates to use workspace metadata references
- Fix unsafe attributes for Rust 2024 edition compatibility
- Move tests from central file to respective modules for better organization
