+++ crates/tree-sitter-tribute/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "tree-sitter-tribute"
-version = "0.1.0"
-edition = "2024"
-authors = ["Tribute Language Team"]
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
 description = "Tree-sitter grammar for the Tribute programming language"
 license = "MIT"
 repository = "https://github.com/your-org/tribute"
