+++ crates/tribute-cranelift/src/tests.rs
@@ -1,6 +1,6 @@
 //! Tests for the Cranelift compiler
 
-use tribute_ast::{parse_source_file, SourceFile, TributeDatabaseImpl};
+use tribute_ast::{SourceFile, TributeDatabaseImpl, parse_source_file};
 use tribute_hir::lower_program_to_hir;
 
 use crate::compile_to_object;
