+++ crates/tribute-hir/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 name = "tribute-hir"
-version = "0.1.0"
-edition = "2024"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
 
 [dependencies]
 tribute-ast.workspace = true
