+++ crates/tribute-runtime/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "tribute-runtime"
-version = "0.1.0"
-edition = "2021"
-authors = ["Tribute Contributors"]
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
 description = "Runtime library for compiled Tribute programs"
 
 [lib]
