+++ crates/tribute-runtime/benches/quick_benchmark.rs
@@ -2,7 +2,7 @@
 //!
 //! Lightweight benchmarks to quickly measure TrString performance
 
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use criterion::{Criterion, black_box, criterion_group, criterion_main};
 use tribute_runtime::memory::*;
 use tribute_runtime::string_ops::*;
