+++ crates/tribute-runtime/src/arithmetic.rs
@@ -4,10 +4,10 @@
 //! type checking and coercion at runtime.
 //! Uses handle-based API for GC compatibility.
 
-use crate::value::{allocation_table, TrHandle, TrValue};
+use crate::value::{TrHandle, TrValue, allocation_table};
 
 /// Add two values - supports number + number and string concatenation
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub extern "C" fn tr_value_add(left: TrHandle, right: TrHandle) -> TrHandle {
     if left.is_null() || right.is_null() {
         return allocation_table().allocate(TrValue::unit());
@@ -49,7 +49,7 @@ pub extern "C" fn tr_value_add(left: TrHandle, right: TrHandle) -> TrHandle {
 }
 
 /// Subtract two values - only supports numbers
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub extern "C" fn tr_value_sub(left: TrHandle, right: TrHandle) -> TrHandle {
     if left.is_null() || right.is_null() {
         return allocation_table().allocate(TrValue::number(0.0));
@@ -75,7 +75,7 @@ pub extern "C" fn tr_value_sub(left: TrHandle, right: TrHandle) -> TrHandle {
 }
 
 /// Multiply two values - only supports numbers
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub extern "C" fn tr_value_mul(left: TrHandle, right: TrHandle) -> TrHandle {
     if left.is_null() || right.is_null() {
         return allocation_table().allocate(TrValue::number(0.0));
@@ -101,7 +101,7 @@ pub extern "C" fn tr_value_mul(left: TrHandle, right: TrHandle) -> TrHandle {
 }
 
 /// Divide two values - only supports numbers
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub extern "C" fn tr_value_div(left: TrHandle, right: TrHandle) -> TrHandle {
     if left.is_null() || right.is_null() {
         return allocation_table().allocate(TrValue::number(0.0));
@@ -129,7 +129,7 @@ pub extern "C" fn tr_value_div(left: TrHandle, right: TrHandle) -> TrHandle {
 }
 
 /// Modulo operation - only supports numbers
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub extern "C" fn tr_value_mod(left: TrHandle, right: TrHandle) -> TrHandle {
     if left.is_null() || right.is_null() {
         return allocation_table().allocate(TrValue::number(0.0));
@@ -155,7 +155,7 @@ pub extern "C" fn tr_value_mod(left: TrHandle, right: TrHandle) -> TrHandle {
 }
 
 /// Negate a value - only supports numbers
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub extern "C" fn tr_value_neg(handle: TrHandle) -> TrHandle {
     if handle.is_null() {
         return allocation_table().allocate(TrValue::number(0.0));
@@ -176,5 +176,63 @@ pub extern "C" fn tr_value_neg(handle: TrHandle) -> TrHandle {
         };
 
         allocation_table().allocate(result)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::memory::*;
+    use crate::value::allocation_table;
+    use serial_test::serial;
+
+    #[test]
+    #[serial]
+    fn test_arithmetic_operations() {
+        // Clear allocation table for test isolation
+        allocation_table().clear();
+        let left = tr_value_from_number(10.0);
+        let right = tr_value_from_number(5.0);
+
+        // Test addition
+        let sum = tr_value_add(left, right);
+        assert_eq!(tr_value_to_number(sum), 15.0);
+        tr_value_free(sum);
+
+        // Test subtraction
+        let diff = tr_value_sub(left, right);
+        assert_eq!(tr_value_to_number(diff), 5.0);
+        tr_value_free(diff);
+
+        // Test multiplication
+        let product = tr_value_mul(left, right);
+        assert_eq!(tr_value_to_number(product), 50.0);
+        tr_value_free(product);
+
+        // Test division
+        let quotient = tr_value_div(left, right);
+        assert_eq!(tr_value_to_number(quotient), 2.0);
+        tr_value_free(quotient);
+
+        // Clean up original values
+        tr_value_free(left);
+        tr_value_free(right);
+    }
+
+    #[test]
+    #[serial]
+    fn test_string_arithmetic() {
+        // Clear allocation table for test isolation
+        allocation_table().clear();
+        let hello = tr_value_from_string("Hello ".as_ptr(), 6);
+        let world = tr_value_from_string("World!".as_ptr(), 6);
+
+        // Test string concatenation through addition
+        let result = tr_value_add(hello, world);
+        assert_eq!(tr_value_get_tag(result), 1); // String tag is 1
+
+        tr_value_free(hello);
+        tr_value_free(world);
+        tr_value_free(result);
     }
 }
